benchmark/GDPval-AA v2/grader/GDPVal_Eval/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gdpval-eval"
version = "0.1.0"
description = "GDPval-AA: Artificial Analysis evaluation framework for OpenAI's GDPval dataset"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.14.1",
    "pandas>=1.5.3",
    "google-generativeai>=0.8.0",
    "requests>=2.31.0",
    "tqdm>=4.64.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["gdpval*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
